+++ frontend/galaxy/GalaxySidebar.tsx
