+++ frontend/galaxy/administration/approvals/Approvals.tsx
@@ -1,6 +1,6 @@
 import { useTranslation } from 'react-i18next';
 import { PageHeader, PageLayout, PageTable } from '../../../../framework';
-import { idKeyFn, useHubView } from '../../useHubView';
+import { idKeyFn, useHubView } from '../../useGalaxyView';
 import { Approval } from './Approval';
 import { useApprovalActions } from './hooks/useApprovalActions';
 import { useApprovalFilters } from './hooks/useApprovalFilters';
