+++ frontend/galaxy/automation-content/collections/hooks/useCollections.tsx
@@ -1,5 +1,5 @@
 import { useGet } from '../../../../common/useItem';
-import { HubItemsResponse } from '../../../useHubView';
+import { HubItemsResponse } from '../../../useGalaxyView';
 import { Collection } from '../Collection';
 
 export function useCollections() {
