+++ frontend/galaxy/automation-content/collections/hooks/useDeleteCollections.tsx
@@ -2,7 +2,7 @@ import { useCallback, useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import { compareStrings, useBulkConfirmation } from '../../../../../framework';
 import { requestDelete } from '../../../../Data';
-import { idKeyFn } from '../../../useHubView';
+import { idKeyFn } from '../../../useGalaxyView';
 import { Collection } from '../Collection';
 import { useCollectionColumns } from './useCollectionColumns';
