+++ frontend/galaxy/useGalaxyView.tsx
